Refactor: moving core recipes away from 'core' package
