+++ src/test/java/org/openrewrite/gitlab/AddComponentTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.gitlab.core;
+package org.openrewrite.gitlab;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
